requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
pandas
joblib
paho-mqtt
requests